[BE] refactor: common type으로 빼둔 state 정의한 enum들 각 도메인으로 이동
