--- server/src/domains/common/enums/AlarmType.ts
@@ -1,8 +1,0 @@
-export enum AlarmType {
-  JOIN_GROUP_REQUEST = 'JOIN_GROUP_REQUEST',
-  MENTORING_REQUEST = 'MENTORING_REQUEST',
-  JOIN_GROUP_ACCEPTED = 'JOIN_GROUP_ACCEPTED',
-  JOIN_GROUP_DECLINED = 'JOIN_GROUP_DECLINED',
-  MENTORING_ACCEPTED = 'MENTORING_ACCEPTED',
-  METTORING_DECLIEND = 'METTORING_DECLIEND',
-}
