--- server/src/domains/common/enums/UserType.ts
@@ -1,5 +1,0 @@
-export enum UserType {
-  USER_MENTOR = 'USER_MENTOR',
-  USER_MENTEE = 'USER_MENTEE',
-  USER_OWNER = 'USER_OWNER',
-}
