--- server/src/domains/common/enums/UsingTechAs.ts
@@ -1,5 +1,0 @@
-export enum UsingTechAs {
-  GROUP = 'GROUP',
-  MENTOR = 'MENTOR',
-  MENTEE = 'MENTEE',
-}
