--- server/src/domains/common/enums/index.ts
@@ -1,4 +1,0 @@
-import { AlarmType } from './AlarmType';
-import { UserType } from './UserType';
-import { UsingTechAs } from './UsingTechAs';
-export { AlarmType, UserType, UsingTechAs };
